Create sine_cosine_waveform_generation
Design a digital frequency synthesis logic in Verilog. The following diagram specifies the interface of the block. Use a sine lookup table (LUT) with 512 entries that stores sin(𝜃!) for 𝑘 = 0,1,…, 511 and 𝜃! = 𝑘2𝜋/512 for the phase in the range of (0, 2𝜋). sin(𝜃!) values stored in the LUT should be fixed point quantized in (10,9) representation (i.e., mantissa values are stored in the LUT). Note that the maximum positive value you can represent in this format is slightly less than 1.0. So you will need to saturate the sin(𝜋/2) to the maximum allowed mantissa value before storing it in the LUT.